% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvcharacteristics.R
\name{cvcharacteristics}
\alias{cvcharacteristics}
\title{cvcharacteristics}
\usage{
cvcharacteristics(fetch = "name", network = NA, park = NA,
  data = "water")
}
\arguments{
\item{fetch}{Indicates what you wish to retrive from the common view data service. Options are:
\describe{
\item{"name"}{Returns the name of the characteristic}
\item{"full"}{Return the:  Charactersiticn name, substrate, sample fraction, category, display name, units, and details.}
\item{"all"}{Returns all data from the "Characteristic" table of the common view.}
}}

\item{network}{A network four letter code in quotes. Only data on parks in that network will be returned.}

\item{park}{A park four letter code in qoutes. Only data from that park will be returned.}

\item{data}{Indicates the type of data you are interesed in. Currently has no function as only water monitoring data can be retrieved.}
}
\value{
Either a \code{vector} with site codes or names, or a \code{data.frame} with all network data.
}
\description{
Returns the name, other data for wat4er qruaily charactersitcs that are avaialable through the NPS common view
}

