% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvdata.R
\name{cvdata}
\alias{cvdata}
\title{cvdata}
\usage{
cvdata(fetch = "all", network = NA, park = NA, site = NA,
  characteristic = NA, data = "water")
}
\arguments{
\item{fetch}{Indicates what you wish to retrive from the common view data service. Options are:
\describe{
\item{"basic"}{Returns the measurement data from the common view.}
\item{"full"}{Return the measurement data as well as metadata on the characteristics and sites.}
\item{"all"}{Returns all data from the Datum, Characteristic", Site, Site, Date and Time tables of the common view.}
}}

\item{network}{A network four letter code in quotes. Only data on parks in that network will be returned.}

\item{park}{A park four letter code in quotes. Only data from that park will be returned.}

\item{site}{A site code in quotes. Only data from that site will be returned.}

\item{characteristic}{A characterisitc name in quotees. Only data from that characteristic will be returned.}

\item{data}{Indicates the type of data you are interesed in. Currently has no function as only water monitoring data can be retrieved.}
}
\value{
A \code{data.frame} with the selected data.
}
\description{
Returns data from the NPS common view.
}

