% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvnetworks.R
\name{cvnetworks}
\alias{cvnetworks}
\title{cvnetworks}
\usage{
cvnetworks(fetch = "code", data = "water")
}
\arguments{
\item{fetch}{Indicates what you wish to retrive from the common view data service. Options are:
\describe{
\item{"code"}{The default. Returns the four letter code of the I & M network}
\item{"name"}{Returns the full name of the I & M network}
 \item{"full"}{Returns both the 4 letter codea and the full name of the I&M network}
 \item{"all"}{Returns all data from the "networks" table of the common view}
}}

\item{data}{Indicates the type of data you are interesed in. Currently has no function as only water monitoring data can be retrieved.}
}
\value{
Eithter a \code{vector} with network codes or names, or a \code{data.frame} with network data.
}
\description{
Returns the name and code of I&M netowrks with data as avaialable through the NPS common view
}

