% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvsites.R
\name{cvsites}
\alias{cvsites}
\title{cvsites}
\usage{
cvsites(fetch = "code", network = NA, park = NA, data = "water")
}
\arguments{
\item{fetch}{Indicates what you wish to retrive from the common view data service. Options are:
\describe{
\item{"code"}{The default. Returns the code for the site}
\item{"name"}{Returns the full name of the site}
\item{"all"}{Returns all data from the "parks" table of the common view as well as the nework name and code.}
}}

\item{network}{A network four letter code in quotes. Only data on parks in that network will be returned.}

\item{park}{A park four letter code in qoutes. Only data from that park will be returned.}

\item{data}{Indicates the type of data you are interesed in. Currently has no function as only water monitoring data can be retrieved.}
}
\value{
Either a \code{vector} with site codes or names, or a \code{data.frame} with all network data.
}
\description{
Returns the name, site code and other data for monitoring sites with data avaialable through the NPS common view
}

