% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterpaste.R
\name{filterpaste}
\alias{filterpaste}
\title{filterpaste}
\usage{
filterpaste(punct = "&", names, values)
}
\arguments{
\item{punct}{The punctuation that should begin the filter arugment. If filter is the first argument it should be "?" otherwise use "&".}

\item{names}{A \code{vector} with the name of the fields you wish to filter on. Names should be in quotes.}

\item{values}{The values for each field indicated in \code{names}. Must be in the same order as \code{names}.}
}
\value{
A text string that can be used for the fiter arguement of an OData call.
}
\description{
Creates the paste argument to be passed to the OData service
}
\details{
This is an internal helper function. It creates the filter options for the OData calls. Currently it cannot handle "or" type arugments. That is, if you want data from two parks - Park 1 or Park2 this cannot create the appropriate filter -  you would have to make 2 data calls. The output of this function is then pasted to the proper url and other arguments to make the complete OData call.
}

