% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filterpaste.R, R/fitlerpaste.R
\name{filterpaste}
\alias{filterpaste}
\title{filterpaste}
\usage{
filterpaste(punct = "&", names, values)

filterpaste(punct = "&", names, values)
}
\arguments{
\item{punct}{The punctuation that should being the filter arugment. If filter is the first argument it should be "?" otherwise use "&".}

\item{names}{A \code{vector} with the name of the fields you wish to filter on. Names shoudl be in quotes}

\item{values}{The values for each field indicated in \code{names}. Must be in the same order as \code{names}.}

\item{punct}{The punctuation that should being the filter arugment. If filter is the first argument it should be "?" otherwise use "&".}

\item{names}{A \code{vector} with the name of the fields you wish to filter on. Names shoudl be in quotes}

\item{values}{The values for each field indicated in \code{names}. Must be in the same order as \code{names}.}
}
\value{
A text string that can be used for the fiter arguement of an OData call.

A text string that can be used for the fiter arguement of an OData call.
}
\description{
Creates the paste arguement to be passed to the OData service

Creates the paste arguement to be passed to the OData service
}
\details{
This is an internal helper function. It creates the filter options for the OData call. Currently it cannot handle "or" type arugments. That is, if you want data from two parks - Park 1 or Park2 this cannot create the apporiate filter -  you would have to make 2 data calls. The output of this fuction is then pasted ot the propoer url and other arugments to make the comnplete Odata call.

This is an internal helper function. It creates the filter options for the OData call. Currently it cannot handle "or" type arugments. That is, if you want data from two parks - Park 1 or Park2 this cannot create the apporiate filter -  you would have to make 2 data calls. The output of this fuction is then pasted ot the propoer url and other arugments to make the comnplete Odata call.
}

